+++ egui_glium/src/backend.rs
@@ -266,7 +266,7 @@ pub fn run(mut app: Box<dyn epi::App>, native_options: epi::NativeOptions) {
                     } else {
                         // Winit uses up all the CPU of one core when returning ControlFlow::Wait.
                         // Sleeping here helps, but still uses 1-3% of CPU :(
-                        if is_focused || !egui.input_state.hovered_files.is_empty() {
+                        if is_focused || !egui.input_state.raw.hovered_files.is_empty() {
                             std::thread::sleep(std::time::Duration::from_millis(10));
                         } else {
                             std::thread::sleep(std::time::Duration::from_millis(50));