+++ egui_glium/src/lib.rs
@@ -47,7 +47,6 @@ pub use copypasta::ClipboardContext; // TODO: remove
 pub struct GliumInputState {
     pub pointer_pos_in_points: Option<Pos2>,
     pub raw: egui::RawInput,
-    pub hovered_files: Vec<std::path::PathBuf>,
 }
 
 impl GliumInputState {
@@ -58,7 +57,6 @@ impl GliumInputState {
                 pixels_per_point: Some(pixels_per_point),
                 ..Default::default()
             },
-            hovered_files: Default::default(),
         }
     }
 }
@@ -257,15 +255,15 @@ pub fn input_to_egui(
                 },
             });
         }
-        WindowEvent::DroppedFile(path) => {
-            input_state.hovered_files.clear();
-            input_state.raw.dropped_files.push(path.clone());
-        }
         WindowEvent::HoveredFile(path) => {
-            input_state.hovered_files.push(path.clone());
+            input_state.raw.hovered_files.push(egui::HoveredFile::from_path(path.clone()));
         }
         WindowEvent::HoveredFileCancelled => {
-            input_state.hovered_files.clear();
+            input_state.raw.hovered_files.clear();
+        }
+        WindowEvent::DroppedFile(path) => {
+            input_state.raw.hovered_files.clear();
+            input_state.raw.dropped_files.push(egui::DroppedFile::from_path(path.clone()));
         }
         _ => {
             // dbg!(event);
@@ -549,7 +547,6 @@ impl EguiGlium {
             None
         };
 
-        self.input_state.raw.hovered_files = self.input_state.hovered_files.clone();
         self.egui_ctx.begin_frame(self.input_state.raw.take());
     }