+++ egui_web/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to the `egui_web` integration will be noted in this file.
 
 ## Unreleased
 
+### Added ⭐
+* Added support for dragging and dropping files into the browser window.
+
 
 ## 0.13.0 - 2021-06-24